ONE TO ONE RELATION ( Character - Weapon )
- CharService = (new DI = httpContextAccessor and GetUserId())
- refactor UserID => GetUserId DI,
- Foreign key, error handling

+++ Services/CharacterService/CharacterService.cs
@@ -1,11 +1,13 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 using AutoMapper;
 using dotnet_rpg.Data;
 using dotnet_rpg.Dtos.Character;
 using dotnet_rpg.Models;
+using Microsoft.AspNetCore.Http;
 using Microsoft.EntityFrameworkCore;
 
 namespace dotnet_rpg.Services.CharacterService
@@ -13,6 +15,7 @@ namespace dotnet_rpg.Services.CharacterService
 	public class CharacterService : ICharacterService
 	{
 		private readonly IMapper _mapper;
+		private IHttpContextAccessor _httpContextAccessor;
 		private readonly DataContext _context;
 
 		private static readonly List<Character> characters = new List<Character>(){
@@ -20,35 +23,47 @@ namespace dotnet_rpg.Services.CharacterService
 			new Character { Id = 1, Name = "Sam"}
 		};
 
-		public CharacterService(IMapper mapper, DataContext context)
+		public CharacterService(IMapper mapper, DataContext context, IHttpContextAccessor httpContextAccessor)
 		{
 			_context = context;
 			_mapper = mapper;
+			_httpContextAccessor = httpContextAccessor;
 		}
 		public async Task<ServiceResponse<List<GetCharacterDto>>> AddCharacter(AddCharacterDto newCharacter)
 		{
 			ServiceResponse<List<GetCharacterDto>> serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
 
 			Character character = _mapper.Map<Character>(newCharacter);
-			character.Id = characters.Max(c => c.Id) +1;
+			character.User = await _context.Users.FirstOrDefaultAsync(u => u.Id == GetUserId());
+			//character.Id = characters.Max(c => c.Id) +1;
 			 
 			await _context.Characters.AddAsync(character);
 			await _context.SaveChangesAsync();
-			serviceResponse.Data = (_context.Characters.Select(c => _mapper.Map<GetCharacterDto>(c))).ToList();
+			serviceResponse.Data = (_context.Characters.Where( c => c.User.Id == GetUserId()).Select(c => _mapper.Map<GetCharacterDto>(c))).ToList();
 
 			return serviceResponse;
 		}
 
+		private int GetUserId() => int.Parse(_httpContextAccessor.HttpContext.User.FindFirstValue(ClaimTypes.NameIdentifier));
+		
 		public async Task<ServiceResponse<List<GetCharacterDto>>> DeleteCharacter(int id)
 		{
 			ServiceResponse<List<GetCharacterDto>> serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
 			try
 			{
-				Character character = await _context.Characters.FirstAsync(c => c.Id == id);
-				_context.Characters.Remove(character);
-				await _context.SaveChangesAsync();
+				Character character = await _context.Characters
+					.FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUserId());
+				if(character != null){
+					_context.Characters.Remove(character);
+					await _context.SaveChangesAsync();
+					serviceResponse.Data = (_context.Characters.Where(c => c.User.Id == GetUserId()).Select(c => _mapper.Map<GetCharacterDto>(c))).ToList();
+				}
+				else
+				{
+					serviceResponse.Success = false;
+					serviceResponse.Message = "Character not found!";
+				}
 
-				serviceResponse.Data = (_context.Characters.Select(c => _mapper.Map<GetCharacterDto>(c))).ToList();
 			}
 			catch (Exception ex)
 			{
@@ -59,10 +74,10 @@ namespace dotnet_rpg.Services.CharacterService
 			return serviceResponse;
 		}
 
-		public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharacters(int userId)
+		public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharacters()
 		{
 			ServiceResponse<List<GetCharacterDto>> serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
-			List<Character> dbCharacters = await _context.Characters.Where(c=>c.User.Id == userId).ToListAsync();
+			List<Character> dbCharacters = await _context.Characters.Where(c=>c.User.Id == GetUserId()).ToListAsync();
 	
 			serviceResponse.Data = (dbCharacters.Select(c => _mapper.Map<GetCharacterDto>(c))).ToList();
 			return serviceResponse;
@@ -71,7 +86,7 @@ namespace dotnet_rpg.Services.CharacterService
 		public async Task<ServiceResponse<GetCharacterDto>> GetCharacterByID(int id)
 		{ 
 			ServiceResponse<GetCharacterDto> serviceResponse = new ServiceResponse<GetCharacterDto>();
-			Character dbCharacter = await _context.Characters.FirstOrDefaultAsync(c => c.Id == id);
+			Character dbCharacter = await _context.Characters.FirstOrDefaultAsync(c => c.Id == GetUserId());
 			serviceResponse.Data = _mapper.Map<GetCharacterDto>(dbCharacter);
 
 			return serviceResponse;
@@ -83,18 +98,25 @@ namespace dotnet_rpg.Services.CharacterService
 
 			try
 			{
-				Character character = await _context.Characters.FirstOrDefaultAsync(c => c.Id == updatedCharacter.Id);
-				character.Name = updatedCharacter.Name;
-				character.Class = updatedCharacter.Class;
-				character.Defence = updatedCharacter.Defence;
-				character.Intelligence = updatedCharacter.Intelligence;
-				character.HitPoints = updatedCharacter.HitPoints;
-				character.Strenght = updatedCharacter.Strenght;
+				Character character = await _context.Characters.Include(c => c.User).FirstOrDefaultAsync(c => c.Id == updatedCharacter.Id);
+				if(character.User.Id == GetUserId())
+				{
+					character.Name = updatedCharacter.Name;
+					character.Class = updatedCharacter.Class;
+					character.Defence = updatedCharacter.Defence;
+					character.Intelligence = updatedCharacter.Intelligence;
+					character.HitPoints = updatedCharacter.HitPoints;
+					character.Strenght = updatedCharacter.Strenght;
 
-				_context.Characters.Update(character);
-				await _context.SaveChangesAsync();
+					_context.Characters.Update(character);
+					await _context.SaveChangesAsync();
 
-				serviceResponse.Data = _mapper.Map<GetCharacterDto>(character);
+					serviceResponse.Data = _mapper.Map<GetCharacterDto>(character);
+				}
+				else{
+					serviceResponse.Success = false;
+					serviceResponse.Message = "character not found";
+				}
 			}
 			catch (Exception ex)
 			{
@@ -102,7 +124,6 @@ namespace dotnet_rpg.Services.CharacterService
 				serviceResponse.Message = ex.Message;
 			}
 			return serviceResponse;
-
 		}
 	}
 }
